requirements.txt
aiopg==1.4.0
anyio==4.2.0
async-timeout==4.0.3
attrdict==2.0.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.95.0
fastapi-sessions==0.3.2
fastapi-utils==0.2.1
future==0.18.3
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.3.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
mergedeep==1.3.4
more-itertools==10.2.0
oauthlib==3.2.2
paramiko==3.4.0
pbr==6.0.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.14
pyjwkest==1.4.2
PyNaCl==1.5.0
pyparsing==3.1.1
requests==2.31.0
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.26
sshtunnel==0.4.0
starlette==0.26.1
testresources==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.21.1
wadllib==1.3.6
zipp==3.17.0
Jinja2==3.1.2
